@@ -5,7 +5,7 @@
     "keywords": [
         "cakephp",
         "php",
-        "pm"
+        "phppm"
     ],
     "homepage": "https://github.com/CakeDC/cakephp-phppm",
     "license": "MIT",
@@ -25,14 +25,11 @@
         "source": "https://github.com/CakeDC/cakephp-phppm"
     },
     "require": {
-        "php": ">=7.2",
-        "cakephp/cakephp": "^4.0",
-        "php-pm/php-pm": "^2.0"
+        "php": ">=8.0",
+        "cakephp/cakephp": "^5.0",
+        "php-pm/php-pm": "^2.3"
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.0",
-        "cakephp/cakephp-codesniffer": "^2.0",
-        "satooshi/php-coveralls": "dev-master"
     },
     "suggest": {
     },
